Rework SSH key handling and configuration
I'm running homu in OpenShift v3/Kubernetes where it's installed (as root) to
`/usr`, but we run as non-root. This follows general best practice that apps
shouldn't be able to mutate their code.

However, we were trying to write the ssh key to `/usr`.  Fix this by
generating a tempfile.  This is also more secure as it closes a prior
race condition where we'd write the file with potentially non-private
access permissions, then chmod it.

Also rework things so that we only write the key once at startup.  By
using `NamedTemporaryFile`, it'll be `unlink()ed` once the object goes
out of scope.  To keep it alive long enough, pass it as an argument to
the "main loop".

Secondly, I want to have the github.com host key in a ConfigMap,
distinct from the code.  And I want to tweak the
ssh configuration for https://stribika.github.io/2015/01/04/secure-secure-shell.html
among other things.

Hence, this also adds a `ssh_config_file` config option.