utils: Print stderr from child processes by default
If for example `git fetch` fails due to a permission error, the github.com ssh
key changing, a segfault inside one of the helper binaries or whatever - I
really want to see that by default.

Currently when running in Kubernetes/Docker I see a Python stack trace, but not
the actual error.